feat：查询所有商品标签

+++ jh-shop-product-service/src/main/java/pers/jhshop/product/model/entity/Categories.java
@@ -16,6 +16,7 @@ import lombok.experimental.Accessors;
 /**
  * <p>
  * 商品分类表
+ * TODO 暂定只会分类到二级
  * </p>
  *
  * @author ChenJiahao(wutiao)
@@ -36,7 +37,8 @@ public class Categories extends Model<Categories> {
 
     @ApiModelProperty(value = "父级分类ID，0表示根分类")
     @TableField("PARENT_ID")
-    private Integer parentId;
+    // TODO 换成Long
+    private Long parentId;
 
     @ApiModelProperty(value = "分类名称(如“手机”，“家电”，“衣物”等)")
     @TableField("NAME")
