+++ jh-shop-product-service/src/main/java/pers/jhshop/product/model/req/CategoriesCreateReq.java
@@ -26,7 +26,7 @@ public class CategoriesCreateReq implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @ApiModelProperty(value = "父级分类ID，0表示根分类")
-    private Integer parentId;
+    private Long parentId;
 
     @ApiModelProperty(value = "分类名称(如“手机”，“家电”，“衣物”等)")
     private String name;
